JPERF-902: Apply axion-release plugin's instructions to CI config
The release CI in the current form requires up to date credentials stored in the Github's secrets feature of the repository.
Tokens are generated for limited amount of time and need to be updated manually.
This requires knowledge maintenance and is doomed to fail at some point.

We are using the credentials (token) in `axion-release` gradle plugin which is dependency of `gradle-release` gradle plugin developed by us.
It is used to create git tags of releases.

With `axion-release` it's possible to reuse github automated mechanisms to authenticate git client operations.
The way to do that is documented in https://axion-release-plugin.readthedocs.io/en/latest/configuration/ci_servers/#github-actions

In order to make the `github.token` work for git client mutable operations like pushing tags we need permission `contents: write` for the CI job.

The approach was inspired by already tested [ci.yml from `report` module](https://github.com/atlassian/report/blob/769530e8acfe14e9b22a2e945d1dedae5a51db38/.github/workflows/ci.yml)

I decided to not grant this permission to the rest of the CI (tests), so I extracted the `release` job.
This way tests kept the default permissions defined on the repository level.